+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed icon handling inside the ChapterRepeatButton.
 - Removed icon handling inside the ChapterToggleButton.
 - Removed icon handling inside the ChapterSelectorBar.
+- Remove option to hide the browse button within the ChapterBrowseTextBox.
 ### Fixed
 - Fixed crash of the ChapterStackPanel if used empty.